Move two more canning subroutines to `oonipl`
