+++ af/shovel/oonipl/const.py
@@ -1,0 +1,6 @@
+#!/usr/bin/env python2.7
+# -*- coding: utf-8 -*-
+
+# https://github.com/ooni/pipeline/blob/master/docs/pipeline-16.10.md#highlights
+# https://gist.github.com/darkk/5736767b8bb1adbf4f5baec24d839d8e#file-bs-txt
+LZ4_LEVEL = '-5'
